backend/requirements.txt
asgiref==3.8.1
Django==5.1.3
django-cors-headers==4.6.0
djangorestframework==3.15.2
pillow==11.0.0
psycopg2==2.9.10
sqlparse==0.5.1
